Preserve user's function
If the user types in a functional expression and then presses a cursor
arrow key, her function is immediately replaced by one of the fixed
functions, as the combo box cycles through its repertoire. Not to worry!
Just scroll the combo back up to its topmost item, and there's her
function, preserved!